requirements.txt
aiohttp==3.9.3
python-binance==1.0.19
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
python-telegram-bot==20.3
ta==0.11.0
colorama==0.4.6
ccxt==4.2.85